[components] Ability to parameterize op name and better default for sling component (#26319)
## Summary & Motivation

Add ability to customize op name.

```yaml
component_type: sling_replication
component_params:
  connections: ...
  op:
    name: my_op
```

In component land, we are going to make it much more clear what are properties on the underlying op versus on the asset, unlike with `@asset`, which commingles it all.

This PR also changes the default op name to be the stem the directory in which it resides.

## How I Tested These Changes

BK