[components] @component and local scope (#26301)
## Summary & Motivation

This PR proposes changing our registration mechanism.

First, instead of an abstract classmethod to determine the registered name, use a decorator `@component`. This plays nicer with inheritance (you do not want to inherit component keys)

`defs.yml` files before this PR had a `component_type` key that indexed into the component library. This works fine for pre-defined components that you don't need to customize. However you often want to specialize execution. In non-component integrations, this means changing the body of an asset-decorated function.

So that end it adds a new component-local syntax fro component type, and reuse that same decorator. The rule is that if the component type is prefixed with a "." we will dynamically load all the python in the folder until a registered component is found.

This works differently. Instead:

* The parent component class defines an `execute` method which has a default implementation
* A child component can override that to customize behavior.

So in the example here, we want to customize the behavior of the sling integration.

```python
@component(name="debug_sling_replication")
class DebugSlingReplicationComponent(SlingReplicationComponent):
    def execute(self, context: AssetExecutionContext, sling: SlingResource) -> Iterator:
        return sling.replicate(context=context, debug=True)
```

And in the `defs.yml` file

```yaml
component_type: .debug_sling_replication
component_params:
  connections:
    - name: snowflake
      type: snowflake
```

The prefix of "." tells the component system to look for a local module instead of going to the component registry.

## How I Tested These Changes

Manual