[dagster-aws][PipesEcsClient] Add support for passing WaiterConfig for ECS pipes client run. (#26266)
Running tasks through ECS pipes is currently limited to jobs that take
under 10 minutes, this is because we rely on the default WaiterConfig
which is currently set to delay of 6 seconds and max attempts of 100,
100*6 = 600 seconds or 10 minutes.

See:
https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ecs/waiter/TasksStopped.html

This commit allows the following parameters to be set when calling
PipesECSClient.run()

* waiter_delay (int)
* waiter_max_attempts (int)

They are optional parameters and will default to `6` and `1000000`
respectively. This equates to a default wait of approx. 70 days. The
reason this was chosen as the default is for two reasons:

1. I think it's a reasonable expectation for people running ECS tasks on
dagster to be able to run beyond 10 minutes.
2. It's what Airflow uses as a default:
https://github.com/apache/airflow/pull/34928

--

## Summary & Motivation

ECS tasks taking longer than 10 minutes would 'detach' on Dagster, I
would rather it did not.

## How I Tested These Changes

See tests added in the PR.

## Changelog

[dagster-aws] `ECSPipesClient.run` now waits up to 70 days for tasks completion (waiter parameters are configurable) 

---------

Co-authored-by: Ian Jenkins <ian.jenkins@liverpoolfc.com>