Check for presence of logStrems in describe_log_streams response.
Sometimes a request to describe_log_streams() can return a valid
response but have no logStreams available. This change handles that by
returning false for target_is_readable() and messages_are_readable() is
the response for describe_log_streams() is valid but length of
logStreams in the response is empty.

It also moves all calls to get_log_events() to a custom function which
implements retry and exponential back off as recommended here: https://repost.aws/knowledge-center/cloudwatch-logs-throttling