[ui] Global AMP: Fix timeline pagination (#26353)
## Summary & Motivation

In the global AMP view, the tick timeline is broken when paginating to older ticks. The (paginated) table view sets the parent's time range, but the time range isn't being used as a signal to refetch the list of ticks used by the timeline. Fix this with an effect.

## How I Tested These Changes

View AMP overview for Discord. Page backward in time, verify that the timeline updates with the correct ticks displayed.

## Changelog

[ui] Fix global auto-materialize tick timeline when paginating.