[ui] Move CodeMirror-based components out of main ui-components export (#26304)
## Summary & Motivation

The CodeMirror-based components in `ui-components` are adding quite a bit of CSS bloat via automatic imports, and the implicit use of `navigator` breaks static rendering in Next.js apps.

Split them out into an `editor` export, which can be used via `@dagster-io/ui-components/lib/editor`.

It would probably be good to split the library up further, but this is at least a step in the right direction.

## How I Tested These Changes

TS, lint, jest. `yarn build` for the component library, verify that it builds correctly and that CM-based components are split out of `index`.