[components] Fix generated path to dbt project (#26384)
## Summary & Motivation

CWD is not what is expected causing the dbt project path to be wrong
when it is calculated as a relative path of the CWD. This PR introduces
a hack with a hardcoded correction. We will fix the root problem of the
CWD in a followup.

## How I Tested These Changes

Manually, observed correct path being generated.