[Selection syntax] Generalized autocomplete functionality (#26273)
## Summary & Motivation

This PR introduces generalized autocomplete functionality that can be
re-used across grammars that follow the base structure of our asset
selection syntax grammar.

As an input the autocomplete factory takes the names of attributes,
their values, and the list of functions available in the grammar.

To implement auto-complete I opted to create a new grammar so that I can
allow certain things to still be valid inputs.
Benefits of separate grammar:

- `IncompleteOrExpression`, `IncompleteAndExpression`,
`IncompleteFunctionCall`, etc. allow our syntax highlighting and
autocomplete visitors to visit the full expression even if parts of it
are malformed.
- DownTraversal and UpTraversal are distinguished simplifying the
visitor logic
- Whitespace is differentiated between `afterExpressionWhitespace` and
`afterLogicalOperatorWhitespace`, etc. greatly simplifying the visitors
that provide suggestions in whitespace. (Instead of needing a visit
function for every possible expression we instead have the whitespace
tokens handle creating suggestions).
- `orToken`, `andToken`, `notToken` simplifies visitor logic, rather
having the various expressions that contain these tokens handle what to
do when the cursor is over them, we can instead have the token itself
handle what happens and reduce the amount of visitor logic.
- Separate tokens for `FunctionName` `AttributeName` `AttributeValue` -
also reduces visitor logic

Generally speaking, my strategy for the grammar was to make leaf nodes
that are shared among the various expression in order to allow those
leaf nodes to make suggestions and reduce the number of visit functions
required in the visitor (avoid combinatorial explosion of cases).

The reason for doing `this.cursorIndex - 1` in the `nodeIncludesCursor`
function is that I want the suggestions to be determined by the token
PRECEEDING the cursor, so I want the token immediately before the cursor
to return "true" for `nodeIncludesCursor`.

I can probably clean up the grammar a bit, but quite fatigued for this
PR. In a future PR I'll look to simplify the grammar even more but what
I have now passes all the tests I've written.

## How I Tested These Changes

Extensive jest testing and real testing.

<img width="937" alt="Screenshot 2024-12-04 at 11 20 58 AM"
src="https://github.com/user-attachments/assets/c29e702e-6048-40cb-ada6-04bb1a1a2a66">