[dagster-snowflake] Add additional_snowflake_connection_args option to SnowflakeResource/IOManager for catchall snowflake connector args (#26040)
## Summary & Motivation
adds an `additional_snowflake_connection_args` config option to the
snowflake resource and i/o manager so that we can support all of the
snowflake connection parameters without needing to add each on to the
config

## How I Tested These Changes

## Changelog

[dagster-snowflake] `SnowflakeResource` and `SnowflakeIOManager` both
accept `additional_snowflake_connection_args` config. This dictionary of
arguments will be passed to the `snowflake.connector.connect` method.
This config will be ignored if you are using the `sqlalchemy` connector.