Handle adding and removing deps in AssetsDefinition.map_asset_specs (#26144)
# Background
This PR changes `AssetsDefinition.map_asset_specs` to allow adding
additional deps to an `AssetSpec`.

To do this, we create `Nothing` inputs under the hood which map to the
asset dependency in question.

This is important mostly for integration cases where we use the
decorator pattern `@dbt_assets`, `@fivetran_assets`, etc. If you want to
link an asset dependency to assets defined with those decorators, it's
currently not feasible for users.

# Caveats

**Only supports op-backed assets**
This PR adds support only for op-backed assets, not graph-backed assets.
I attempted to add support and found myself blocked since graph-backed
assets currently do not support non-argument deps (at least from what I
could gather). I added a linear issue to track this and referenced it in
the relevant place in code.

**Doesn't work for argument-based deps**
This PR adds support for the deletion of non-argument based deps; but
does not support the deletion of argument-based deps. I think deletion
of argument-based deps is impossible, since you would have nothing to
pass to the underlying function. We error cleanly instead, and I think
that's the best we can do.

# Testing

A bunch of tests for every case that I could think of:
- add additional dependencies
- add the same dependency multiple times
- remove a dependency
- add dependencies with partition mappings
- add dependencies to a non-executable assets def
- attempt to add deps to a graph-backed asset