[components] Add support for extras (conditional imports version) (#26440)
## Summary & Motivation

Add support for extras in component discovery in `dagster-components`.
Alternative to #26417 (conditional entry point loading) that instead
uses conditional import statements:

- Discovery mechanism has changed so that submodules of the target
module of an entry point are no longer traversed. Instead we look only
at the specified module for declared components.
- Because we load everything coming from the entry point target, and we
unconditionally load every declared entry point, we place the burden on
component library authors to examine the environment and skip loading
any components corresponding to uninstalled extras.
- Implementation of this approach in `dagster_components.lib`, where we
condition imports of components for the `dbt` and `sling` extras on the
presence of `dagster-dbt` and `dagster-embedded-elt` packages in the
environment. Basically we are just hardcoding in
`dagster_components/lib/__init__.py` a repeat of the dependencies
associated with each extra rather than reading them from package
metadata (approach in #26417).

## How I Tested These Changes

Unit tests confirming only the correct components are registered when
each extra is installed.