[components] Simplify component folder creation (#26438)
## Summary & Motivation

There were very confusion calls to `generate_project` and `generate_subtree` in our scaffolding code that did nothing but generate a folder, even though they referred to nonexistent templates. 

Turns out this worked because `os.walk` silent returns an empty generator when it points to a directory that does not exist.

Just changing this to a `mkdir_p` call for clarity and simplicity.

## How I Tested These Changes

Manual test. BK