[bugfix] k8s client now waits for main container to be ready (#26422)
## Summary & Motivation

A more complete fix for issue #23497. The recent change in PR #24432
exits the loop if all sidecar init containers are in a ready state, and
doesn't wait until the regular (non init container). If the main
(dagster) container takes longer to startup, for example if it's using a
large image, the code fails when it attempts to read the log.

## How I Tested These Changes
* Fixed the existing unit tests to represent the full state of the pod
when we expect it to be ready. The existing tests did not ensure that
the main container is ready, and instead returned when the init
containers were ready (or terminated)
* Tested locally with our k8s deployment and deployed to production to
fix the issue for our jobs

## Changelog

> Insert changelog entry or delete this section.