[freshness-checks][rfc] Retrieve latest completed check execution in summary records. (#26331)
## Summary & Motivation
This is in preparation to make some fixes to the freshness check sensor,
which needs to understand both when a check is in flight, and the last
completed execution of the check.

Previously, on an AssetCheckExecutionSummaryRecord, we were only storing
the latest record regardless of status. For freshness checks, we
additionally need to know the latest _completed_ execution of a check.
This PR adds that behavior by making it possible to query check
execution history based on status; a cloud PR companion will follow up
if we think this is reasonable.

Here's my intuition regarding perf:
- In most cases, this will retrieve the latest 2 check executions at
most.
- In some degenerate cases, we could potentially have to scan many check
execution records if there are many in flight checks at once for the
same check key. But it's hard to imagine this happening reasonably
often.

## How I Tested These Changes
Added new assertions to event log storage tests, a new assertion in
instance tests to show what the status behavior is for runless events.