[freshness-checks][rfc] fix tickets (#26335)
## Summary & Motivation
Attempts to fix tickets for freshness checks:
https://linear.app/dagster-labs/issue/BUILD-527/%5Bfreshness-sensor%5D-check-for-newer-materialization-when-last-check
and
https://linear.app/dagster-labs/issue/BUILD-528/%5Bfreshness-sensor%5D-catch-long-runtime-issues-even-when-checks-are#comment-8cfd22cb.

The exact cases in question:
- When the asset becomes overdue because the evaluation is taking too
long, and there is a planned freshness check evaluation as part of the
run, we should still kick off a run of the check so that users can get
alerted.
- When there's a materialization more recent than the latest check
result, kick off another evaluation of the check.

## How I Tested These Changes
I added new tests for the previously failing cases.

## Changelog
[bugfixes]
- Previously, the freshness check sensor would not re-evaluate freshness
checks if an in-flight run was planning on evaluating that check. Now,
the freshness check sensor will kick off an independent run of the
check, even if there's already an in flight run, as long as the
freshness check can potentially fail.
- Previously, if the freshness check was in a failing state, the sensor
would wait for a run of some sort to update the freshness check before
re-evaluating. Now, if there's a materialization later than the last
evaluation of the freshness check and no planned evaluation, we will
re-evaluate the freshness check automatically.