[components] Stop forcing editable dagster for `dg generate code-location` (#26447)
## Summary & Motivation

Removes the forced use of an editable installation for
`dagster-dg`-generated code locations. Allows for dependencies to be
added incrementally.

`dg generate code-location` takes a `--use-editable-dagster` option.
This can either be used as a flag, in which case env var
`DAGSTER_GIT_REPO_DIR` will be read, or takes a value pointing to the
directory containing the dagster repo.

## How I Tested These Changes

Unit tests and ran through demo using published code on this branch.