speed up computing a graphql id for a GraphenePipelineSnapshot
Summary:
Right now the graphql ID is the same as the snapshot ID, but the snapshot ID can be slow to compute for large jobs. Use a faster ID instead when you don't already know the snapshot ID, and pass in the snapshot ID instead when you do. Audited existing callsites of GraphenePipelineSelector to ensure that they don't expect the id to match the selector ID.

> Insert changelog entry or delete this section.