[components] Refactor generate files to call explicit function to build component.yaml (#26439)
## Summary & Motivation

I'm in the process of creating scaffolding for generic custom components and want to customize the scaffolding process more. This means giving more explicit control to users to define their own scaffolding to some degree.

Hence I'm making it so that the user explicitly invokes the creation of a component.yaml file rather than having our framework do it. This lets the user customize the behavior. I also think this is more obvious and understandable, at the cost of an extra import.

In order to ensure that a custom component author doesn't screw this up, we add a check after `generate_files` to ensure that `component.yaml` exists. 

## How I Tested These Changes

BK and manual