[components] Add dagster-dg global config and AppContext (#26516)
## Summary & Motivation

Add a global `AppContext` and `DgConfig` object in `dagster-dg`. The
purpose of these abstractions is to capture global configuration and
pass it throughout the system. Global configuration is supported via
top-level options passed to `dg` on the command line.

`AppContext` may appear premature in this PR, since it just wraps
`DgConfig`. Upstack we add the cache to it.

## How I Tested These Changes

Unit tests.