[components] Rendering engine (#26468)
## Summary & Motivation

This is the initial layer for a jinja-based rendering engine for the contents of `components.yaml`. To keep this diff simpler, I didn't hook this up to anythin quite yet.

The key bit here is that in some cases, we want to defer the templating logic until a point in time where we will have more context available. For example, the asset translation logic for the dbt_project component needs to apply its template to each individual node, which won't be available until runtime.

Other fields need to be substituted in before the initial load, such as those being passed into resources used during the component init process.

This allows us to distinguish between these two cases by setting metadata in the generated json schema, which is then used to determine if rendering should be deferred or not.

Note that rendering is done at a per-value level, which caps the amount of craziness someone can do in these (i.e. you can't dynamically generate a nested dictionary of fields). This is generally intentional, I think once you get to that level of complexity you should really be doing that in python land, and attempting to support that sort of thing would make the mental model a lot more complicated.

## How I Tested These Changes

## Changelog

NOCHANGELOG