feat: allow passing config for dagster asset materialize cli [#26069 continued] (#26578)
## Summary & Motivation

Continuation of #26069:

> I was browsing the repo for good first issues as a new contributor,
and found #16802 interesting. It looks like @sam-goodwin's PR #20979 is
nearly complete, except for a change needed for one of the imports. I
used that PR's solution, but also resolved the import issue.

Which is a continuation of #20979:

> > This change adds support for -c, --config and --config-json in the
dagster asset materialize CLI command.
> > Our motivation is to simply materialize a whole graph as part of
CI/CD and use run_config to specify parameters that scope down that
materialization for the sake of CI/CD. A single partition of our asset
graph can take a long time to run and for CI/CD we don't need the whole
thing.

## How I Tested These Changes

## Changelog

- [dagster-cli] Adds support for `--config-json` to the `dagster asset
materialize` command

---------

Co-authored-by: Jimmy Shah <jimmybshah01@gmail.com>