[module-loaders] Genericize object list functionality to take in all sensors, jobs, schedules (#26545)
## Summary & Motivation
Actually genericize `ModuleScopedDagsterObjects` to handle taking in all
types of Dagster objects - sensors, schedules, and jobs. I explicitly
left out resources and loggers, because I don't think it makes sense to
scoop those up at module load - since they need to be associated with a
key.

## How I Tested These Changes
Added a new test file `test_module_loaders` which scaffolds out a module
fake for a given test spec, and ensures that either an error is thrown
or the correct number of attributes are accessible on the underlying
list.