add support for incremental step stats calculations (#26577)
## Summary & Motivation
This is the counterpart of
https://github.com/dagster-io/dagster/pull/26550, but for step stats
instead of run stats.

The step stats calculations are a little more complicated because it
relies on intermediate state in between dagster events to be held. This
includes the `previous_attempt_start` time for a given step. This PR
introduces a field in the snapshot to carry over this state, which
should only be used in the incremental calculation.

Another change made here is that marker events (e.g. resource
init/teardown) can be associated with a step, but generally occur before
the step has started. That means that previously step snapshots did not
include these marker events if the step had not yet started. This PR
changes that behavior so that every step will have a snapshot if it has
a marker event, even if it has not started. The output for incremental
vs completed run snapshots should match exactly.

## How I Tested These Changes
BK