[components] Add --rebuild-component-registry option, `--skip-venv` option, build cache on code loc construction (#26591)
## Summary & Motivation

Add a top-level `--rebuild-component-registry` option that clears the
cache and refetches the component registry for the current environment.
The reason this is called `--rebuild-component-registry` instead of
`--rebuild-cache` is:

- It is scoped to just the current environment, whereas the cache
encompasses multiple environments
- We may cache other info later

Also add `--skip-venv` option to `dg generate code-location` to avoid
automatic venv setup and cache population on construction. Mostly useful
for testing.

We also invoke this immediately after `uv sync` when constructing a new
code location, so that the cache is loaded.

## How I Tested These Changes

New unit tests + manual test of code location generation.