[partition-metadata][rfc] add_asset_metadata (#26562)
## Summary & Motivation
Introduce an `AssetExecutionContext.add_asset_metadata` method, which
will add metadata to materialization events (but not output events).

Storing metadata only on asset events seems like a naturally good thing
to do. It's primarily what we display in the UI / is primarily what is
queryable by users, and not storing it on output events saves db write
size.

For partitioned assets, you are able to specify per-partition metadata
that will only be added to the relevant partitioned materialization.
This solves a long-standing user request to have per-partition handling
in `add_output_metadata` more naturally.

## How I Tested These Changes
Added a bunch of new tests for both non-partitioned and partitioned
cases.

## Changelog
A new `AssetExecutionContext.add_asset_metadata` method, which allows
you to add metadata to materialization events for a given asset or asset
partition. See the API docs here:
https://docs.dagster.io/_modules/dagster/_core/execution/context/asset_execution_context#AssetExecutionContext.add_asset_metadata