Add weakMapMemoize utility. (#26605)
## Summary & Motivation

Introduces a memoize function thats more powerful than what lodash
offers. Lodash memoize only works with string keys, and so it requires
you to take your arguments and somehow serialize it to a string.
Depending on the object this can be very slow if the object is big. To
avoid the headache of serialization I'm adding a weakMapMemoize which is
much more flexible and can cache any function mixing both primitives and
objects.

Going to use this to memoize the `filterByQuery` helpers for the
GanttChart since it looks like we call the utility in a lot of different
places separately and having them coordinate so that there's only 1 call
is a bit tricky.


## How I Tested These Changes

jest tests