Make UpstreamPartitionsResult subset-aware (#26044)
Summary:
This allows us to make the method on AssetGraphView
compute_parent_subset_and_required_but_nonexistent_subset instead of
still working in terms of a list of partition keys.

I'm a little trepeditious about this because I could imagine cases where
the invalid upstream partition is so invalid that it can't even be
expressed as part of a PartitionsSubset? I guess in that case we could
just raise an exception when you try to do the mapping like we already
do in many cases (e.g. the timezones not matching).

The trickiest bit here was making sure that it was still possible to
create and pass around an 'invalid' partitions subset (since we are a
bit inconsistent in our APIs about whether or not we allow that when
doing the various transformations from keys to subset and back).

Test Plan: BK

NOCHANGELOG

## Summary & Motivation

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.