[13/n][dagster-airbyte] Implement airbyte_assets and build_airbyte_assets_definitions (#26432)
## Summary & Motivation

This PR implements the `airbyte_assets` decorator and the
`build_airbyte_assets_definitions` factory.
- `airbyte_assets` can be used to create all assets for a given Airbyte
connection, i.e. one asset per table in the connection.
- `build_airbyte_assets_definitions` can be used to create all Airbyte
assets defs, one per connection. It uses `airbyte_assets`.

These can be used with `AirbyteCloudWorkspace`, and the goal would be to
use these with the workspace for Airbyte OSS as well.

## How I Tested These Changes

Additional unit tests with BK.

## Changelog

[dagster-airbyte] The `airbyte_assets` decorator is added. It can be
used with the `AirbyteCloudWorkspace` resource and
`DagsterAirbyteTranslator` translator to load Airbyte tables for a given
connection as assets in Dagster. The `build_airbyte_assets_definitions`
factory can be used to create assets for all the connections in your
Airbyte workspace.