[dagster-airlift] Handle run retries (#25761)
## Summary & Motivation
Previously, when launching runs to a "migrated" asset, we would ignore
run retries - so if the original run failed, regardless of whether a
retry succeeded, we would fail the airflow run.

This PR fixes to correctly account for run retries in Dagster - by
hooking into the retry machinery that @jamiedemaria built.

## How I Tested These Changes
Adds a new test which simulates having auto-run-execution by using
`materialize` - and succeeding the last retry. Ensure that the airflow
run succeeds, and the asset materializes eventually.

The reason it's necessary to simulate auto-run-execution is because we
use sqlite in tests as opposed to postgres - which we need in order to
enable run retries.

I initially attempted to materialize assets inline within the body of
the asset itself, but that seems to screw up some global context, so
instead do it via graphql queries.

## Changelog
- [dagster-airlift] correctly support auto-run-retries when kicking of
materializations of Dagster-proxied assets from airflow.