[ui] Update DA evaluation page to show backfill in table if a backfill was requested (#26643)
## Summary & Motivation
The runs table for a DA evaluation was not showing backfills when DA
requested a backfill
<img width="1440" alt="Screenshot 2024-12-20 at 10 55 57 AM"
src="https://github.com/user-attachments/assets/e38e5598-4e35-4219-a0de-9228b99d595e"
/>

This is because the the `selectedEvaluation.runIds` could be run ids or
a backfill id, but we were always adding them as run id filters. Then
the value was a backfill id, it would return no results.

This add some (admittedly a bit hacky) logic to determine if the id is a
backfill id, and set the correct filter if so. A more long term solution
would be to store run ids and backfill ids as separate lists on the
evaluation, but this will fix the problem until we can make the more
sustainable change.


<img width="1440" alt="Screenshot 2024-12-20 at 11 18 30 AM"
src="https://github.com/user-attachments/assets/2541bb18-b011-49d9-b271-56f5f022cf64"
/>

I also had to update the `RunsFeedTableWithFilters` component to take a
bool `includeRunsFromBackfills` so that we would only show the backfill
row, not the runs that the backfill launched. I had to then update all
of the users of `RunsFeedTableWithFilters` to pass the bool

## How I Tested These Changes
👀 

## Changelog

[ui] Fixed a bug where backfills launched by Declarative Automation were
not being shown in the table of launched runs.